Move xml native impl clases to builtin package
