+++ modules/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/builtin/xmllib/Copy.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,7 +16,7 @@
  * under the License.
  **/
 
-package org.ballerinalang.nativeimpl.lang.xmls;
+package org.ballerinalang.nativeimpl.builtin.xmllib;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
@@ -24,36 +24,34 @@ import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.model.values.BXML;
 import org.ballerinalang.nativeimpl.lang.utils.ErrorHandler;
 import org.ballerinalang.natives.AbstractNativeFunction;
-import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
+import org.ballerinalang.natives.annotations.ReturnType;
 
 /**
- * Remove an attribute from an XML.
- * 
- * @since 0.95
+ * Make a deep copy of an XML.
  */
 @BallerinaFunction(
         packageName = "ballerina.builtin",
-        functionName = "xml.removeAttribute",
-        args = {@Argument(name = "qname", type = TypeKind.STRING)},
+        functionName = "xml.copy",
+        returnType = {@ReturnType(type = TypeKind.XML)},
         isPublic = true
 )
-public class RemoveAttribute extends AbstractNativeFunction {
+public class Copy extends AbstractNativeFunction {
 
-    private static final String OPERATION = "remove attribute";
+private static final String OPERATION = "get children from xml";
 
     @Override
     public BValue[] execute(Context ctx) {
+        BValue result = null;
         try {
             // Accessing Parameters.
-            BXML<?> xml = (BXML<?>) getRefArgument(ctx, 0);
-            String qname = getStringArgument(ctx, 0);
-            xml.removeAttribute(qname);
+            BXML value = (BXML) getRefArgument(ctx, 0);
+            result = value.copy();
         } catch (Throwable e) {
             ErrorHandler.handleXMLException(OPERATION, e);
         }
-        
+
         // Setting output value.
-        return VOID_RETURN;
+        return getBValues(result);
     }
 }
