+++ modules/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/builtin/xmllib/GetTextValue.java
@@ -16,45 +16,44 @@
  * under the License.
  **/
 
-package org.ballerinalang.nativeimpl.lang.xmls;
+package org.ballerinalang.nativeimpl.builtin.xmllib;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BMap;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.model.values.BXML;
 import org.ballerinalang.nativeimpl.lang.utils.ErrorHandler;
 import org.ballerinalang.natives.AbstractNativeFunction;
-import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
+import org.ballerinalang.natives.annotations.ReturnType;
 
 /**
- * Sets the attributes the provided attributes map
+ * Get the text value of a XML.
  * 
- * @since 0.90
+ * @since 0.88
  */
 @BallerinaFunction(
         packageName = "ballerina.builtin",
-        functionName = "xml.setAttributes",
-        args = {@Argument(name = "attributes", type = TypeKind.MAP)},
+        functionName = "xml.getTextValue",
+        returnType = {@ReturnType(type = TypeKind.STRING)},
         isPublic = true
 )
-public class SetAttributes extends AbstractNativeFunction {
+public class GetTextValue extends AbstractNativeFunction {
 
-    private static final String OPERATION = "set attributes";
+    private static final String OPERATION = "get text from xml";
 
     @Override
     public BValue[] execute(Context ctx) {
+        BValue result = null;
         try {
             // Accessing Parameters.
-            BXML<?> xml = (BXML<?>) getRefArgument(ctx, 0);
-            BMap<String, ?> attributes = (BMap<String, ?>) getRefArgument(ctx, 1);
-            xml.setAttributes(attributes);
+            BXML xml = (BXML) getRefArgument(ctx, 0);
+                result = xml.getTextValue();
         } catch (Throwable e) {
             ErrorHandler.handleXMLException(OPERATION, e);
         }
         
         // Setting output value.
-        return VOID_RETURN;
+        return getBValues(result);
     }
 }
