+++ modules/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/builtin/xmllib/IsEmpty.java
@@ -16,7 +16,7 @@
  * under the License.
  **/
 
-package org.ballerinalang.nativeimpl.lang.xmls;
+package org.ballerinalang.nativeimpl.builtin.xmllib;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
@@ -24,38 +24,35 @@ import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.model.values.BXML;
 import org.ballerinalang.nativeimpl.lang.utils.ErrorHandler;
 import org.ballerinalang.natives.AbstractNativeFunction;
-import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
 
 /**
- * Get all the elements-type items in the given sequence, that matches a given qualified name.
+ * Check whether the XML sequence is empty.
  * 
  * @since 0.88
  */
 @BallerinaFunction(
         packageName = "ballerina.builtin",
-        functionName = "xml.select",
-        args = {@Argument(name = "qname", type = TypeKind.STRING)},
-        returnType = {@ReturnType(type = TypeKind.XML)},
+        functionName = "xml.isEmpty",
+        returnType = {@ReturnType(type = TypeKind.BOOLEAN)},
         isPublic = true
 )
-public class Select extends AbstractNativeFunction {
+public class IsEmpty extends AbstractNativeFunction {
 
-    private static final String OPERATION = "select elements from xml";
+    private static final String OPERATION = "check xml is empty";
 
     @Override
     public BValue[] execute(Context ctx) {
         BValue result = null;
         try {
             // Accessing Parameters.
-            BXML value = (BXML) getRefArgument(ctx, 0);
-            String qname = getStringArgument(ctx, 0);
-            result = value.elements(qname);
+            BXML xml = (BXML) getRefArgument(ctx, 0);
+            result = xml.isEmpty();
         } catch (Throwable e) {
             ErrorHandler.handleXMLException(OPERATION, e);
         }
-
+        
         // Setting output value.
         return getBValues(result);
     }
