+++ modules/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/builtin/xmllib/Select.java
@@ -16,7 +16,7 @@
  * under the License.
  **/
 
-package org.ballerinalang.nativeimpl.lang.xmls;
+package org.ballerinalang.nativeimpl.builtin.xmllib;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
@@ -29,34 +29,33 @@ import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
 
 /**
- * Searches in children recursively for elements matching the name and returns a sequence containing them all.
- * Does not search within a matched result.
+ * Get all the elements-type items in the given sequence, that matches a given qualified name.
  * 
- * @since 0.92
+ * @since 0.88
  */
 @BallerinaFunction(
         packageName = "ballerina.builtin",
-        functionName = "xml.selectDescendants",
+        functionName = "xml.select",
         args = {@Argument(name = "qname", type = TypeKind.STRING)},
         returnType = {@ReturnType(type = TypeKind.XML)},
         isPublic = true
 )
-public class SelectDescendants extends AbstractNativeFunction {
+public class Select extends AbstractNativeFunction {
 
-    private static final String OPERATION = "select descendants from xml";
+    private static final String OPERATION = "select elements from xml";
 
     @Override
     public BValue[] execute(Context ctx) {
         BValue result = null;
         try {
             // Accessing Parameters.
-            BXML<?> value = (BXML<?>) getRefArgument(ctx, 0);
+            BXML value = (BXML) getRefArgument(ctx, 0);
             String qname = getStringArgument(ctx, 0);
-            result = value.descendants(qname);
+            result = value.elements(qname);
         } catch (Throwable e) {
             ErrorHandler.handleXMLException(OPERATION, e);
         }
-        
+
         // Setting output value.
         return getBValues(result);
     }
