+++ modules/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/builtin/xmllib/SelectChildren.java
@@ -16,7 +16,7 @@
  * under the License.
  **/
 
-package org.ballerinalang.nativeimpl.lang.xmls;
+package org.ballerinalang.nativeimpl.builtin.xmllib;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
@@ -24,23 +24,25 @@ import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.model.values.BXML;
 import org.ballerinalang.nativeimpl.lang.utils.ErrorHandler;
 import org.ballerinalang.natives.AbstractNativeFunction;
+import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
 
 /**
- * Get all the elements-type items of a xml.
+ * Selects and concatenate all the children of the elements in this sequence that matches the given qualified name.
  * 
  * @since 0.88
  */
 @BallerinaFunction(
         packageName = "ballerina.builtin",
-        functionName = "xml.elements",
+        functionName = "xml.selectChildren",
+        args = {@Argument(name = "qname", type = TypeKind.STRING)},
         returnType = {@ReturnType(type = TypeKind.XML)},
         isPublic = true
 )
-public class Elements extends AbstractNativeFunction {
+public class SelectChildren extends AbstractNativeFunction {
 
-    private static final String OPERATION = "get elements from xml";
+    private static final String OPERATION = "select children from xml";
 
     @Override
     public BValue[] execute(Context ctx) {
@@ -48,7 +50,8 @@ public class Elements extends AbstractNativeFunction {
         try {
             // Accessing Parameters.
             BXML value = (BXML) getRefArgument(ctx, 0);
-            result = value.elements();
+            String qname = getStringArgument(ctx, 0);
+            result = value.children(qname);
         } catch (Throwable e) {
             ErrorHandler.handleXMLException(OPERATION, e);
         }
